move snakemake_nersc_cli to own module
